actualiacion de tabla pagos

+++ database/migrations/2023_03_21_131236_create_pagos_table.php
@@ -21,7 +21,7 @@ class CreatePagosTable extends Migration
             $table->string('num_comprobante');
             $table->date('fecha');
             $table->binary('imagen_comprobante');
-            $table->integer('precio');
+            $table->float('precio');
             $table->timestamps();
         });
     }
